refactor: add useLoginFetch custom hook and use in LoginPage.tsx
